Separate mock and real steps for AuthManager
